Fixed test 0007, added test 0014, improved error handling
Test 0007 --exit-status when this opt is set and the last value returned
is null or false we set exit-status specially

Test 0014 Line numbers reported for errors should now match the user
script. Column does also, but it's the parenthesized number.

Overall errors are handled internally with more of them being bubbled up
to the top